requirements.txt
streamlit==1.28.1
PyMuPDF==1.23.8
pdfplumber==0.10.3
openai==2.14.0
python-dotenv==1.0.0
chromadb==0.4.14
langchain==0.1.1
reportlab==4.0.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pinecone-client==2.2.2
pytest==7.4.3
tiktoken==0.5.2